sorting and precision fixes for tests

+++ src/openms/include/OpenMS/ANALYSIS/ID/BasicProteinInferenceAlgorithm.h
@@ -137,7 +137,7 @@ namespace OpenMS
     double getInitScoreForAggMethod_(const AggregationMethod& aggregation_method, bool higher_better) const;
 
     /// get lambda function to aggregate scores
-    typedef float (*fptr)(float, float);
+    typedef double (*fptr)(double, double);
     fptr aggFunFromEnum_(const BasicProteinInferenceAlgorithm::AggregationMethod& agg_method, bool higher_better) const;
   };
 } //namespace OpenMS
